+++ packages/http-client-csharp/src/lib/client-model-builder.ts
@@ -29,24 +29,23 @@ import {
 import { getVersions } from "@typespec/versioning";
 import { $lib } from "../emitter.js";
 import { NetEmitterOptions, resolveOptions } from "../options.js";
-import { ClientKind } from "../type/clientKind.js";
-import { CodeModel } from "../type/codeModel.js";
-import { InputClient } from "../type/inputClient.js";
-import { InputConstant } from "../type/inputConstant.js";
-import { InputOperation } from "../type/inputOperation.js";
-import { InputOperationParameterKind } from "../type/inputOperationParameterKind.js";
-import { InputParameter } from "../type/inputParameter.js";
-import { InputPrimitiveTypeKind } from "../type/inputPrimitiveTypeKind.js";
-import { InputEnumType, InputModelType, InputPrimitiveType } from "../type/inputType.js";
-import { InputTypeKind } from "../type/inputTypeKind.js";
-import { RequestLocation } from "../type/requestLocation.js";
+import { ClientKind } from "../type/client-kind.js";
+import { CodeModel } from "../type/code-model.js";
+import { InputClient } from "../type/input-client.js";
+import { InputConstant } from "../type/input-constant.js";
+import { InputOperationParameterKind } from "../type/input-operation-parameter-kind.js";
+import { InputOperation } from "../type/input-operation.js";
+import { InputParameter } from "../type/input-parameter.js";
+import { InputPrimitiveTypeKind } from "../type/input-primitive-type-kind.js";
+import { InputTypeKind } from "../type/input-type-kind.js";
+import { InputEnumType, InputModelType, InputPrimitiveType } from "../type/input-type.js";
+import { RequestLocation } from "../type/request-location.js";
 import { Usage } from "../type/usage.js";
-import { getExternalDocs } from "./decorators.js";
 import { logger } from "./logger.js";
 import { getUsages, navigateModels } from "./model.js";
 import { loadOperation } from "./operation.js";
-import { processServiceAuthentication } from "./serviceAuthentication.js";
-import { resolveServers } from "./typespecServer.js";
+import { processServiceAuthentication } from "./service-authentication.js";
+import { resolveServers } from "./typespec-server.js";
 import { createContentTypeOrAcceptParameter } from "./utils.js";
 
 export function createModel(sdkContext: SdkContext<NetEmitterOptions>): CodeModel {
@@ -71,7 +70,6 @@ export function createModelForService(
   sdkContext: SdkContext<NetEmitterOptions>,
   service: Service
 ): CodeModel {
-  const emitterOptions = resolveOptions(sdkContext.emitContext);
   const program = sdkContext.emitContext.program;
   const serviceNamespaceType = service.type;
 
@@ -96,7 +94,6 @@ export function createModelForService(
     : undefined;
 
   const description = getDoc(program, serviceNamespaceType);
-  const externalDocs = getExternalDocs(sdkContext, serviceNamespaceType);
 
   const servers = getServers(program, serviceNamespaceType);
   const namespace = getNamespaceFullName(serviceNamespaceType) || "client";
