+++ packages/http-client-csharp/src/type/body-media-type.ts
