+++ packages/http-client-csharp/src/type/client-kind.ts
