+++ packages/http-client-csharp/src/type/code-model.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputAuth } from "./inputAuth.js";
-import { InputClient } from "./inputClient.js";
-import { InputEnumType, InputModelType } from "./inputType.js";
+import { InputAuth } from "./input-auth.js";
+import { InputClient } from "./input-client.js";
+import { InputEnumType, InputModelType } from "./input-type.js";
 
 export interface CodeModel {
   Name: string;
