+++ packages/http-client-csharp/src/type/collection-format.ts
