+++ packages/http-client-csharp/src/type/external-docs.ts
