+++ packages/http-client-csharp/src/type/formatted-type.ts
