+++ packages/http-client-csharp/src/type/http-response-header.ts
@@ -1,14 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputType } from "./inputType.js";
+import { InputType } from "./input-type.js";
 
-export interface InputModelProperty {
+export interface HttpResponseHeader {
   Name: string;
-  SerializedName: string;
+  NameInResponse: string;
   Description: string;
   Type: InputType;
-  IsRequired: boolean;
-  IsReadOnly: boolean;
-  IsDiscriminator?: boolean;
 }
