+++ packages/http-client-csharp/src/type/input-api-key-auth.ts
