+++ packages/http-client-csharp/src/type/input-auth.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+import { InputOAuth2Auth } from "./input-oauth2-auth.js";
 import { InputApiKeyAuth } from "./inputApiKeyAuth.js";
-import { InputOAuth2Auth } from "./inputOAuth2Auth.js";
 
 export interface InputAuth {
   ApiKey?: InputApiKeyAuth;
