+++ packages/http-client-csharp/src/type/input-constant.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputType } from "./inputType.js";
+import { InputType } from "./input-type.js";
 
 export interface InputConstant {
   Value?: any;
