+++ packages/http-client-csharp/src/type/input-enum-type-value.ts
