+++ packages/http-client-csharp/src/type/input-intrinsic-type-kind.ts
