+++ packages/http-client-csharp/src/type/input-model-property.ts
@@ -1,11 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputType } from "./inputType.js";
+import { InputType } from "./input-type.js";
 
-export interface HttpResponseHeader {
+export interface InputModelProperty {
   Name: string;
-  NameInResponse: string;
+  SerializedName: string;
   Description: string;
   Type: InputType;
+  IsRequired: boolean;
+  IsReadOnly: boolean;
+  IsDiscriminator?: boolean;
 }
