+++ packages/http-client-csharp/src/type/input-oauth2-auth.ts
