+++ packages/http-client-csharp/src/type/input-operation-parameter-kind.ts
