+++ packages/http-client-csharp/src/type/input-primitive-type-kind.ts
