+++ packages/http-client-csharp/src/type/input-type-kind.ts
