+++ packages/http-client-csharp/src/type/input-type-serialization-format.ts
