+++ packages/http-client-csharp/src/type/input-type-value.ts
