+++ packages/http-client-csharp/src/type/literal-type-context.ts
