+++ packages/http-client-csharp/src/type/request-location.ts
