+++ packages/http-client-csharp/test/Unit/encode.test.ts
@@ -3,16 +3,16 @@ import { getAllHttpServices } from "@typespec/http";
 import assert, { deepStrictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
 import { loadOperation } from "../../src/lib/operation.js";
-import { InputPrimitiveTypeKind } from "../../src/type/inputPrimitiveTypeKind.js";
-import { InputEnumType, InputModelType, InputPrimitiveType } from "../../src/type/inputType.js";
-import { InputTypeKind } from "../../src/type/inputTypeKind.js";
+import { InputPrimitiveTypeKind } from "../../src/type/input-primitive-type-kind.js";
+import { InputTypeKind } from "../../src/type/input-type-kind.js";
+import { InputEnumType, InputModelType, InputPrimitiveType } from "../../src/type/input-type.js";
 import {
   createEmitterContext,
   createEmitterTestHost,
   createNetSdkContext,
   navigateModels,
   typeSpecCompile,
-} from "./utils/TestUtil.js";
+} from "./utils/test-util.js";
 
 describe("Test encode duration", () => {
   let runner: TestHost;
@@ -155,36 +155,7 @@ describe("Test encode duration", () => {
     );
   });
 
-  it("encode iso8601 on duration model property", async () => {
-    const program = await typeSpecCompile(
-      `
-            @doc("This is a model.")
-            model ISO8601DurationProperty {
-                @encode(DurationKnownEncoding.ISO8601)
-                value: duration;
-            }
-      `,
-      runner
-    );
-    const context = createEmitterContext(program);
-    const sdkContext = createNetSdkContext(context);
-    const [services] = getAllHttpServices(program);
-    const modelMap = new Map<string, InputModelType>();
-    const enumMap = new Map<string, InputEnumType>();
-    navigateModels(sdkContext, services[0].namespace, modelMap, enumMap);
-    const durationProperty = modelMap.get("ISO8601DurationProperty");
-    assert(durationProperty !== undefined);
-    deepStrictEqual(
-      {
-        Kind: InputTypeKind.Primitive,
-        Name: InputPrimitiveTypeKind.DurationISO8601,
-        IsNullable: false,
-      } as InputPrimitiveType,
-      durationProperty.Properties[0].Type
-    );
-  });
-
-  it("encode seconds-int32 on duration model property", async () => {
+  it("encode seconds int32 on duration model property", async () => {
     const program = await typeSpecCompile(
       `
             @doc("This is a model.")
@@ -213,7 +184,7 @@ describe("Test encode duration", () => {
     );
   });
 
-  it("encode seconds-int32 on duration model property", async () => {
+  it("encode seconds float on duration model property", async () => {
     const program = await typeSpecCompile(
       `
             @doc("This is a model.")
