+++ packages/http-client-csharp/test/Unit/property-type.test.ts
@@ -1,16 +1,16 @@
 import { TestHost } from "@typespec/compiler/testing";
 import assert, { deepStrictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
-import { createModel } from "../../src/lib/clientModelBuilder.js";
-import { InputPrimitiveTypeKind } from "../../src/type/inputPrimitiveTypeKind.js";
-import { InputEnumType, InputListType } from "../../src/type/inputType.js";
-import { InputTypeKind } from "../../src/type/inputTypeKind.js";
+import { createModel } from "../../src/lib/client-model-builder.js";
+import { InputPrimitiveTypeKind } from "../../src/type/input-primitive-type-kind.js";
+import { InputTypeKind } from "../../src/type/input-type-kind.js";
+import { InputEnumType, InputListType } from "../../src/type/input-type.js";
 import {
   createEmitterContext,
   createEmitterTestHost,
   createNetSdkContext,
   typeSpecCompile,
-} from "./utils/TestUtil.js";
+} from "./utils/test-util.js";
 
 describe("Test GetInputType for array", () => {
   let runner: TestHost;
