+++ packages/http-client-csharp/test/Unit/string-format.test.ts
@@ -3,16 +3,16 @@ import { getAllHttpServices } from "@typespec/http";
 import assert, { deepStrictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
 import { loadOperation } from "../../src/lib/operation.js";
-import { InputPrimitiveTypeKind } from "../../src/type/inputPrimitiveTypeKind.js";
-import { InputEnumType, InputModelType, InputPrimitiveType } from "../../src/type/inputType.js";
-import { InputTypeKind } from "../../src/type/inputTypeKind.js";
+import { InputPrimitiveTypeKind } from "../../src/type/input-primitive-type-kind.js";
+import { InputTypeKind } from "../../src/type/input-type-kind.js";
+import { InputEnumType, InputModelType, InputPrimitiveType } from "../../src/type/input-type.js";
 import {
   createEmitterContext,
   createEmitterTestHost,
   createNetSdkContext,
   navigateModels,
   typeSpecCompile,
-} from "./utils/TestUtil.js";
+} from "./utils/test-util.js";
 
 describe("Test string format", () => {
   let runner: TestHost;
