+++ packages/http-client-csharp/test/Unit/usage.test.ts
@@ -8,7 +8,7 @@ import {
   createEmitterTestHost,
   createNetSdkContext,
   typeSpecCompile,
-} from "./utils/TestUtil.js";
+} from "./utils/test-util.js";
 
 describe("Test getUsages", () => {
   let runner: TestHost;
