+++ packages/http-client-csharp/test/Unit/utils/test-util.ts
@@ -16,7 +16,7 @@ import { RestTestLibrary } from "@typespec/rest/testing";
 import { VersioningTestLibrary } from "@typespec/versioning/testing";
 import { getFormattedType, getInputType } from "../../../src/lib/model.js";
 import { NetEmitterOptions } from "../../../src/options.js";
-import { InputEnumType, InputModelType } from "../../../src/type/inputType.js";
+import { InputEnumType, InputModelType } from "../../../src/type/input-type.js";
 
 export async function createEmitterTestHost(): Promise<TestHost> {
   return createTestHost({
